회원 정보 조회, 수정

+++ src/main/java/com/kidsqueue/kidsqueue/parent/db/ParentRepository.java
@@ -2,7 +2,7 @@ package com.kidsqueue.kidsqueue.parent.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParentRepository extends JpaRepository<Parent, Integer> {
+public interface ParentRepository extends JpaRepository<Parent, Long> {
 
     Boolean existsByLoginId(String loginId);
